requirements.txt
ipython
openai
python-dotenv
streamlit
